services/auth/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
prometheus-fastapi-instrumentator==6.1.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
